requirements.txt
django
django-phonenumber-field
phonenumbers
pillow
django-debug-toolbar